fix: resolve all C4244/C4267 integer conversion warnings on x64
Widen the tip window timer IDs to UINT_PTR to match SetTimer, and add
explicit narrowing casts where LRESULT, LPARAM, INT_PTR, ptrdiff_t and
size_t values feed int/UINT/COLORREF sinks. All builds are warning-free
on Win32 and x64.

+++ SpaceMonger.cpp
@@ -128,7 +128,7 @@ void CSpaceMonger::OnAppExit(void)
 void CSpaceMonger::OnFileOpen(void)
 {
 	CDriveDialog drvdialog;
-	int drive = drvdialog.DoModal();
+	int drive = (int)drvdialog.DoModal();
 	if (drive == -1) return;
 
 	CFolderTree *ft = (CFolderTree *)m_document;
